Further modularization of input processing
