+++ core/src/com/fourbit/pc_invader/PcInvader.java
@@ -3,13 +3,10 @@ package com.fourbit.pc_invader;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.ApplicationAdapter;
 import com.badlogic.gdx.graphics.GL20;
-import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.graphics.Pixmap;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.graphics.glutils.FrameBuffer;
-import com.badlogic.gdx.math.Vector2;
-import com.badlogic.gdx.math.Vector3;
 import com.badlogic.gdx.utils.ScreenUtils;
 import com.fourbit.pc_invader.levels.BossLevel;
 import com.fourbit.pc_invader.levels.Level;
@@ -77,20 +74,5 @@ public class PcInvader extends ApplicationAdapter {
         this.level.dispose();
         this.batch.dispose();
         this.sceneFrameBuffer.dispose();
-    }
-
-    public static Vector2 getMouseCoords() {
-        Vector2 mouseCoords2D = new Vector2();
-        Vector3 mouseCoords3D = new Vector3();
-        OrthographicCamera cam = new OrthographicCamera();
-
-        mouseCoords3D.x = Gdx.input.getX();
-        mouseCoords3D.y = GAME_HEIGHT - Gdx.input.getY();
-        mouseCoords3D.z = 0;
-        cam.unproject(mouseCoords3D);
-        mouseCoords2D.x = mouseCoords3D.x;
-        mouseCoords2D.y = mouseCoords3D.y;
-
-        return mouseCoords2D;
     }
 }
